Version 2.0.0 (Aided patient performance prediction)
Version 1.2.0 and 2.0.0 introduce new features which were used to
perform simulation for listeners with impaired hearing using hearing
aids [1].

The main new features are:

1) A new FADE module called "processing"

It can be used prior to the training stage to process the corpus wav
files with some signal procesing. A short tutorial was added, too.
Two examples configurations for procressing are provided:
a) "fade.d/processing.d/copy" simply copies the files and can be easily
adapted to use any programm to perform the processing in parallel.
b) "fade.d/processing.d/openMHA" is an example based on the open Master
Hearing aid (openMHA.org) and requires the openMHA software to be
installed and added to the PATH environment variable. This example can
be easily extended with arbirary openMHA configurations.
The default behavior is to skip this module.

2) A modified version of the SGBFB features implementing audiograms

This feature extraction (fade/fade.d/features.d/sgbfb-abel) loads
audiograms from "load_audiogram.m" and uses them to eliminate the signal
portions below the hearing thresholds in the log Mel-spectrogram.
Because the left and right feature channels are calculated separately
and then simply concatenated, we called it automatic better ear
listening (ABEL).

In addition, the number of model states was increased from six to eigth.
This change improves the SRTs in fluctuating noise conditions for some
languages of the Matrix test and is not expected to interfere with other
results (e.g. tone in noise detection). A corresponding publication is
currently prepared.

Another new, but barely tested, feature is the possibility to add
impulse responses to the source folder which are then used to process
the speech and noise signals with the corresponding responses during
corpus generation. Please don't rely on that feature yet.

[1] https://doi.org/10.1177/2331216518768954