Add code for audio quality predictions
This feature was developed with the aim to predict audio quality with
FADE.
It was never officially evaluated and published.
Time ago, I tried to predict the ranks of the codecs in one of the OPUS
codec comparions (I think it was the 96kbit-one).
It worked, but I never got around to finish this work.
The idea was to run the training on unprocessed samples, and the
recognition on processed samples.
The loglikelihood of the (of course correctly recognized tokens) was
used as a measure of similarity.